requirements.txt
django==5.0
setuptools==68.2.0
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf_yasg==1.21.7
faker==20.1.0
django-cors-headers==4.3.1
